15차 수정(결재 승인,반려 기능 완료 및 쓸데없는 파일 삭제)

+++ src/main/java/org/zerock/chain/parkyeongmin/repository/ApprovalRepository.java
@@ -5,9 +5,10 @@ import org.springframework.stereotype.Repository;
 import org.zerock.chain.parkyeongmin.model.Approval;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
-public interface ApprovalRepository extends JpaRepository<Approval, Integer> {
+public interface ApprovalRepository extends JpaRepository<Approval, Long> {
 
     // 특정 결재자의 모든 문서를 docNo로 최신순으로 조회
     List<Approval> findByEmployeeEmpNoOrderByDocumentsDocNoDesc(Long empNo);
